Update ReadMe file
The examples in the readme files used some configuration extension methods that would not compile.  For example,  'AddServiceBusCommandDispatcher' and 'AddServiceBusEventPublisher'.  These have been replaced with 'AddServiceBus' which works as the examples intend in a sample project.

Other minor changes spacing and formatting adjustments have been made, as suggested by a markdown linter.